+++ pkg/mls/store/models.go
@@ -1,4 +1,4 @@
-package mlsstore
+package store
 
 import "github.com/uptrace/bun"
