+++ pkg/topic/topic.go
@@ -40,9 +40,9 @@ func Category(contentTopic string) string {
 }
 
 func BuildGroupTopic(groupId string) string {
-	return fmt.Sprintf("/xmtp/3/g-%s/proto", groupId)
+	return fmt.Sprintf("/xmtp/mls/1/g-%s/proto", groupId)
 }
 
 func BuildWelcomeTopic(installationId []byte) string {
-	return fmt.Sprintf("/xmtp/3/w-%x/proto", installationId)
+	return fmt.Sprintf("/xmtp/mls/1/w-%x/proto", installationId)
 }
